+++ CHANGELOG.md
@@ -10,6 +10,8 @@ about writing changes to this log.
 
 ## [Unreleased]
 
+- Added application tests.
+
 ## [1.1.2] 2023-02-17
 
 - [SUPP0RT-874](https://jira.itkdev.dk/browse/SUPP0RT-874)